+++ tests/autoclassifier/test_image.py
@@ -21,8 +21,8 @@ def test_fit():
     model = AutoImageClassifier(
         datamodule,
         max_epochs=1,
-        timeout=10,
+        timeout=5,
         suggested_backbones="ssl_resnet18",
         n_trials=1,
     )
-    model.fit()
+    model.hp_tune()
